% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_TV.R
\name{ci_TV}
\alias{ci_TV}
\title{Construct confidence interval for total variation}
\usage{
ci_TV(
  df,
  d,
  m,
  y,
  at_group,
  ordering = NULL,
  B = 500,
  alpha = 0.05,
  weight.matrix = "diag"
)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable, which is assumed to take a discrete
support}

\item{at_group}{Indicates for which always-taker group the confidence
interval is calculated for. It can take any value in the support of M.}

\item{ordering}{A list with length equal to the cardinality of the support of
the mediator variable. The name of each element corresponds to a point in
the support, and each element is a vector that collects all m values that
are less than or equal to this point. If ordering = NULL, the standard
ordering is used.}

\item{B}{Bootstrap size, default is zero}

\item{alpha}{Significance level}

\item{weight.matrix}{Weight matrix used to implement FSST. Possible options
are "diag", "avar", "identity." Defaults is "diag" as in FSST.}

\item{k}{Indicates which al}
}
\description{
This function calculates a confidence interval for the total
  variation for a given always taker group. The inference is via inverting
  the test of Fang et al. (2023; FSST)
}
