% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tv_ats.R
\name{compute_min_defiers_binaryM}
\alias{compute_min_defiers_binaryM}
\title{Computes the breakdown number of defiers consistent with the sharp null
when M is binary}
\usage{
compute_min_defiers_binaryM(
  df,
  d,
  m,
  y,
  w = NULL,
  continuous_Y = base::ifelse(is.null(num_Ybins), TRUE, FALSE),
  num_Ybins = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df. Must be binary}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable}

\item{w}{(Optional) Name of weighting variable. If null, equal weights are used}

\item{continuous_Y}{(Optional) Whether Y should be treated as continuous, in which case kernel density is used, or discrete. Default is TRUE.}

\item{num_Ybins}{(Optional) If specified, Y is discretized into the given number of bins (if num_Ybins is larger than the number of unique values of Y, no changes are made)}
}
\description{
Computes the minimum fraction of the population that would need
to be defiers to be consistent with the sharp null.
Also computes the minimum ratio of defiers to compliers
}
