% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tv_ats.R
\name{compute_tv_ats}
\alias{compute_tv_ats}
\title{Bounds on fraction of always-takers affected}
\usage{
compute_tv_ats(
  df,
  d,
  m,
  y,
  w = NULL,
  continuous_Y = base::ifelse(is.null(num_Ybins), TRUE, FALSE),
  num_Ybins = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable}

\item{w}{(Optional) Name of weighting variable. If null, equal weights are used}

\item{continuous_Y}{(Optional) Whether Y should be treated as continuous, in which case kernel density is used, or discrete. Default is TRUE.}

\item{num_Ybins}{(Optional) If specified, Y is discretized into the given number of bins (if num_Ybins is larger than the number of unique values of Y, no changes are made)}
}
\description{
Computes bounds on the TV distance btwn Y(1,1) and Y(1,0) for always takers
This is a lower bound on the fraction of ATs for whom there is a direct effect of D on Y
}
