% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_frac_affected.R
\name{compute_tv_ats_multiple_m}
\alias{compute_tv_ats_multiple_m}
\title{Wrapper for lb_frac_affected for legacy issues}
\usage{
compute_tv_ats_multiple_m(...)
}
\description{
Wrapper for lb_frac_affected for legacy issues
}
