% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_shi_nonuisance.R
\name{cox_shi_nonuisance}
\alias{cox_shi_nonuisance}
\title{Run the Cox and Shi test for the null E[Y] <= 0 for Y~N(mu,Sigma)}
\usage{
cox_shi_nonuisance(Y, sigma, alpha = 0.05, refinement = FALSE, tol = 1e-08)
}
\arguments{
\item{Y}{}

\item{sigma}{}

\item{alpha}{}

\item{refinement}{}
}
\description{
Run the Cox and Shi test for the null E[Y] <= 0 for Y~N(mu,Sigma)
}
