% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_y.R
\name{discretize_y}
\alias{discretize_y}
\title{Function for discretizing yvec.}
\usage{
discretize_y(yvec, numBins)
}
\arguments{
\item{yvec}{A vector containing y values (assumed non-missing)}

\item{numBins}{The target number of bins into which to discretize y}
}
\value{
A discretized version of yvec. If yvec has fewer than numBins unique
 values, then yvec is returned. Otherwise, create cutpoints for y using the
 1/numBins, 2/numBins.... quantiles of y. If two cutpoints are the same (owing to point mass)
 we group those together into one bin, so the number of bins returns may be less than numBins
}
\description{
Function for discretizing yvec.
}
