% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sharp_null.R
\name{get_beta.obs}
\alias{get_beta.obs}
\title{Function to construct beta.obs to pass to lpinfer}
\usage{
get_beta.obs(yvec, dvec, mvec)
}
\arguments{
\item{yvec}{Vector of outcome variable}

\item{dvec}{Vector of treatment variable}

\item{mvec}{Vector of mediator variable}
}
\description{
Function to construct beta.obs to pass to lpinfer
}
