% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sharp_null.R, R/test_sharp_null_fsst.R
\name{get_beta.obs_alt}
\alias{get_beta.obs_alt}
\title{Function to construct beta.obs to pass to lpinfer under an alternative
  LP formulation}
\usage{
get_beta.obs_alt(yvec, dvec, mvec)

get_beta.obs_alt(yvec, dvec, mvec)
}
\arguments{
\item{yvec}{Vector of outcome variable}

\item{dvec}{Vector of treatment variable}

\item{mvec}{Vector of mediator variable}
}
\description{
Function to construct beta.obs to pass to lpinfer under an alternative
  LP formulation

Function to construct beta.obs to pass to lpinfer under an alternative
  LP formulation
}
