% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_frac_affected.R
\name{lb_frac_affected}
\alias{lb_frac_affected}
\title{Bounds on fraction of always-takers affected with a multi-valued
  discrete or multi-dimensional discrete m}
\usage{
lb_frac_affected(
  df,
  d,
  m,
  y,
  at_group = NULL,
  w = NULL,
  continuous_Y = base::ifelse(is.null(num_Ybins), TRUE, FALSE),
  num_Ybins = NULL,
  max_defiers_share = 0,
  allow_min_defiers = FALSE,
  return_min_defiers = FALSE
)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Vector of the mediator variable}

\item{y}{Name of the outcome variable}

\item{at_group}{(Optional) Value of m specifying which always-takers to
compute lower bounds of TV for.  If at_group is specified, then we compute
a lower bound on TV between Y(1,at_group) and Y(0,at_group) for ATs who
have M(1)=M(0)=at_group. If at_group is null (the default), we compute a
lower bound on the weighted average of TV across all always-takers, with
weights proportional to shares in population}

\item{w}{(Optional) Name of weighting variable. If null, equal weights are
used}

\item{continuous_Y}{(Optional) Whether Y should be treated as continuous, in
which case kernel density is used, or discrete. Default is TRUE.}

\item{num_Ybins}{(Optional) If specified, Y is discretized into the given
number of bins (if num_Ybins is larger than the number of unique values of
Y, no changes are made)}

\item{max_defiers_share}{(Optional) If specified, up to max_defier_share
fraction of the population can be defiers (of any type). Default is zero}

\item{allow_min_defiers}{(Optional) If the bound on defiers
(max_defier_share) is inconsistent with the data, proceed by allowing the
minimum number of defiers compatible with the data. Otherwise, throw an
error. Default is FALSE}

\item{return_min_defiers}{(Optional) If true, the function returns the
minimum number of defiers consistent with the data rather than the TV
bounds}
}
\description{
Computes a lower bound on the TV distance btwn Y(1,m) and Y(0,m)
  for always takers who have M(1)=M(0)=m if at_group=m. This is a lower bound
  on the fraction of these ATs for whom there is a direct effect of D on Y If
  at_group=NULL, it computes a lower bound on the weighted average across all
  ATs, weighted by their proportion in the population
}
