% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_density_plot.R
\name{partial_density_plot}
\alias{partial_density_plot}
\title{Create plots of partial densities}
\usage{
partial_density_plot(df, d, m, y, numGridPoints = 10000)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable}

\item{numGridPoints}{Number of points used in grid for graphing}
}
\value{
A ggplot object showing partial densities
}
\description{
Plots f_{Y,M=1 | D=1} and f_{Y,M=1 | D=0}
}
