% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_missing_from_df.R
\name{remove_missing_from_df}
\alias{remove_missing_from_df}
\title{This function processes the inputted dataframe to remove missing values}
\usage{
remove_missing_from_df(df, d, m, y, w = NULL)
}
\arguments{
\item{df}{A data.frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable}

\item{w}{(Optional) Name of weighting variable. If null, equal weights are used}
}
\description{
This function processes the inputted dataframe to remove missing values
}
