% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_frac_affected.R
\name{Rglpk_solve_fractional_LP}
\alias{Rglpk_solve_fractional_LP}
\title{Wrapper for Rglpk_solve_lp that implements linear fractional programming}
\usage{
Rglpk_solve_fractional_LP(
  obj_numerator,
  obj_denominator,
  mat,
  dir,
  rhs,
  bounds,
  max = FALSE,
  constant_numerator = 0,
  constant_denominator = 0,
  ...
)
}
\arguments{
\item{obj_numerator}{The objective vector for the numerator of the linear fractional program}

\item{obj_denominator}{The objective vector for the denominator of the linear fractional program. Note that the denominator must be strictly positive over all feasible values (if not, we return a warning and NaN for the value)}

\item{mat}{The constraint matrix}

\item{dir}{The direction of the constraints}

\item{rhs}{The right hand side of the constraints}

\item{bounds}{The bounds on the optimization vector. Note: the bounds should only be 0 or Inf, otherwise this will mess up the transformation of the fractional-LP to an LP. If you have other bounds, add these to the constraints in mat}

\item{max}{(Optional) Whether to maximize or minimize the objective}

\item{constant_numerator}{(Optional) A constant to add to the numerator of the linear fractional program}

\item{constant_denominator}{(Optional) A constant to add to the denominator of the linear fractional program}

\item{...}{(Optional) Additional arguments to be passed to Rglpk_solve_lp}
}
\value{
The optimal value of the linear fractional program
}
\description{
Wrapper for Rglpk_solve_lp that implements linear fractional programming
}
