% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tv_ats_multiple_m.R
\name{row_equals}
\alias{row_equals}
\title{Function that computes whether a row of a df x equals a row of a df y in all elements}
\usage{
row_equals(x, y)
}
\description{
Function that computes whether a row of a df x equals a row of a df y in all elements
}
