% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sharp_null.R
\name{test_sharp_null}
\alias{test_sharp_null}
\title{Hypothesis test for the sharp null of full mediation}
\usage{
test_sharp_null(
  df,
  d,
  m,
  y,
  method,
  ordering = NULL,
  B = 500,
  cluster = NULL,
  weight.matrix = "diag",
  hybrid_kappa = alpha/10,
  num_Ybins = NULL,
  alpha = 0.05,
  rearrange = FALSE,
  eps_bar = 0.001,
  enumerate = FALSE,
  fix_n1 = TRUE,
  lambda = "dd",
  use_nc = FALSE,
  analytic_variance = FALSE,
  defiers_share = 0
)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable(s)}

\item{y}{Name of the outcome variable, which is assumed to take a discrete
support}

\item{method}{Method to use. One of "ARP, "CS", "FSST", "CR"}

\item{ordering}{A list with length equal to the cardinality of the support of the mediator variable. The name of each element corresponds to a point in the support, and each element is a vector that collects all m values that are less than or equal to this point. If ordering = NULL, the standard ordering is used. If length(m) > 1, then the default is the elementwise ordering.}

\item{B}{Bootstrap size, default is 500}

\item{cluster}{Cluster for bootstrap}

\item{weight.matrix}{Weight matrix used to implement FSST. Possible options are "diag", "avar", "identity." Defaults is "diag" as in FSST.}

\item{hybrid_kappa}{The first-stage size value of the ARP hybrid test. If NULL, the ARP conditional (non-hybrid) test is used. Default is alpha/10}

\item{num_Ybins}{(Optional) If specified, Y is discretized into the given number of bins (if num_Ybins is larger than the number of unique values of Y, no changes are made)}

\item{alpha}{Significance level. Default is 0.05}

\item{rearrange}{Logical variable indicating whether to rearrange the conditional probabilities to obey monotonicity. De}

\item{eps_bar}{Cho and Russell (2023) truncation parameter}

\item{enumerate}{Enumerate vertices for Cox and Shi (2023) implementataion?}

\item{fix_n1}{Whether the number of treated units (or clusters) should be fixed in the bootstrap}

\item{lambda}{Lambda value for FSST. Default is "dd" in which case the "data-driven" lambda recommended by FSST is used. If lambda = "ndd", the non data-driven recommendation is used. See Remark 4.2 of FSST.}

\item{use_nc}{If the data is clustered, should FSST use the number of clusters for determining lambda (versus total observations). Default is false.}

\item{analytic_variance}{If TRUE, we use the analytic formula for the variance, rather than a bootstrap. Available if method if ARP or CS. Default is FALSE}

\item{defiers_share}{Bound on the proportion of defiers in the population. Default is 0 which indicates that the monotonicity constraint is imposed.}
}
\description{
This function tests the sharp null that Y(1,m) = Y(0,m). The
  outcome and mediator are both assumed to take finitely many different
  values. Several options are available for infernece.
}
