% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R, R/test_sharp_null_cr.R
\name{test_sharp_null_cr}
\alias{test_sharp_null_cr}
\title{Hypothesis test for the sharp null via Cho and Russell (2023; CR)}
\usage{
test_sharp_null_cr(
  df,
  d,
  m,
  y,
  ordering = NULL,
  B = 500,
  eps_bar = 0.001,
  alpha = 0.05
)

test_sharp_null_cr(
  df,
  d,
  m,
  y,
  ordering = NULL,
  B = 500,
  eps_bar = 0.001,
  alpha = 0.05
)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable, which is assumed to take a discrete
support}

\item{ordering}{A list with length equal to the cardinality of the support of
the mediator variable. The name of each element corresponds to a point in
the support, and each element is a vector that collects all m values that
are less than or equal to this point. If ordering = NULL, the standard
ordering is used.}

\item{B}{Bootstrap size, default is 500}

\item{eps_bar}{Perturbation parameter used to perturb the
objective/constraints}

\item{alpha}{Significance level. Default value is .05}
}
\description{
This function tests the sharp null of Y(1,m) = Y(0,m). The
  outcome and mediator are both assumed to take finitely many different
  values. The inference is via applying CR

This function tests the sharp null of Y(1,m) = Y(0,m). The
  outcome and mediator are both assumed to take finitely many different
  values. The inference is via applying CR
}
