% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sharp_null_toru.R
\name{test_sharp_null_toru}
\alias{test_sharp_null_toru}
\title{Hypothesis test for the sharp null via Kitagawa (2015)}
\usage{
test_sharp_null_toru(
  df,
  d,
  m,
  y,
  B = 500,
  alpha = 0.05,
  num_Ybins = NULL,
  cluster = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable, which is assumed to take a discrete
support}

\item{B}{Bootstrap size, default is 500}

\item{alpha}{Significance level. Default value is .05}

\item{num_Ybins}{(Optional) If specified, Y is discretized into the given number of bins (if num_Ybins is larger than the number of unique values of Y, no changes are made)}
}
\description{
This function tests the sharp null of Y(1,m) = Y(0,m). The
  mediator must be discrete.
}
