% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bounds_ats.R
\name{compute_bounds_ats}
\alias{compute_bounds_ats}
\title{Bounds on direct for always-takers}
\usage{
compute_bounds_ats(df, d, m, y)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable}
}
\description{
Computes bounds on E[Y(1,1) - Y(1,0) | G = AT], the average direct effect for ATs for whom there is a direct effect of D on Y
}
