% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bounds_ats.R
\name{compute_bounds_ats}
\alias{compute_bounds_ats}
\title{Bounds on direct effect for always-takers}
\usage{
compute_bounds_ats(
  df,
  d,
  m,
  y,
  w = NULL,
  c_at_ratio = NULL,
  adjust_for_point_mass = TRUE,
  num_Ybins = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable}

\item{w}{(Optional) Name of weighting variable. If null, equal weights are used}

\item{c_at_ratio}{(Optional) specify the ratio of E[Y(1,1) | C]/E[Y(1,1) | AT]. If this is specified, direct effect for ATs is point-identified}

\item{adjust_for_point_mass}{(Optional) specify whether to use a correction to the bounds that allows for point-mass in the distribution of Y (default TRUE)}

\item{num_Ybins}{(Optional) If specified, Y is discretized into the given number of bins (if num_Ybins is larger than the number of unique values of Y, no changes are made)}
}
\description{
Computes bounds on E[Y(1,1) - Y(1,0) | G = AT], the average direct effect for ATs for whom there is a direct effect of D on Y
}
