% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tv_ats.R
\name{compute_tv_ats}
\alias{compute_tv_ats}
\title{Bounds on fraction of always-takers effected}
\usage{
compute_tv_ats(df, d, m, y, method = "density")
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable}
}
\description{
Computes bounds on the TV distance btwn Y(1,1) and Y(1,0) for always takers
This is a lower bound on the fraction of ATs for whom there is a direct effect of D on Y
}
