% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tv_ats_multiple_m.R
\name{compute_tv_ats_multiple_m}
\alias{compute_tv_ats_multiple_m}
\title{Bounds on fraction of always-takers affected with a multi-valued discrete or multi-dimensional discrete m}
\usage{
compute_tv_ats_multiple_m(df, d, m, y, at_group = NULL, w = NULL)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Vector of the mediator variable}

\item{y}{Name of the outcome variable}

\item{at_group}{(Optional) Value of m specifying which always-takers to compute lower bounds of TV for.
If at_group is specified, then we compute a lower bound on TV between Y(1,at_group) and Y(0,at_group) for
ATs who have M(1)=M(0)=at_group. If at_group is null (the default), we compute a lower bound on
the weighted average of TV across all always-takers, with weights proportional to shares in population}

\item{w}{(Optional) Name of weighting variable. If null, equal weights are used}
}
\description{
Computes  a lower bound on the TV distance btwn Y(1,m) and Y(0,m) for always takers
who have M(1)=M(0)=m if at_group=m. This is a lower bound on the fraction of these ATs for whom there is a direct effect of D on Y
If at_group=NULL, it computes a lower bound on the weighted average across all ATs, weighted by their proportion in the population
}
