% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_density_plot.R
\name{partial_density_plot}
\alias{partial_density_plot}
\title{Create plots of partial densities}
\usage{
partial_density_plot(
  df,
  d,
  m,
  y,
  numGridPoints = 10000,
  plot_nts = FALSE,
  density_1_label = "f(Y,M=1|D=1)",
  density_0_label = "f(Y,M=1|D=0)",
  num_Ybins = NULL,
  continuous_Y = base::ifelse(is.null(num_Ybins), TRUE, FALSE)
)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable}

\item{numGridPoints}{Number of points used in grid for graphing}

\item{plot_nts}{(Optional) If TRUE, we plot f_{Y,M=0 | D=1} and f_{Y,M=0 | D=0}. Otherwise, we plot f_{Y,M=1 | D=1} and f_{Y,M=1 | D=0}. Default is FALSE}

\item{density_1_label}{(Optional) The label on the plot for the d=1 density.}

\item{density_0_label}{(Optional) The label on the plot for the d=0 density.}

\item{num_Ybins}{(Optional) If specified, Y is discretized into the given number of bins (if num_Ybins is larger than the number of unique values of Y, no changes are made)}

\item{continuous_Y}{(Optional) Should Y be treated as continuous for density estimation. Default is TRUE. Use FALSE for discrete Y}
}
\value{
A ggplot object showing partial densities
}
\description{
Plots f_{Y,M=m | D=1} and f_{Y,M=m | D=0} for m equal to 0 or 1
}
